requirements.txt
click==6.6
cookies==2.2.1
dnspython==1.14.0
funcsigs==1.0.2
mock==2.0.0
pbr==1.10.0
py==1.10.0
pytest==3.0.1
requests==2.20.0
responses==0.5.1
six==1.10.0
SQLAlchemy==1.3.11
